% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_management.R
\name{create_analysis_config}
\alias{create_analysis_config}
\title{Create a new configuration file template}
\usage{
create_analysis_config(output_path, template_type = "generic")
}
\arguments{
\item{output_path}{Where to save the configuration file}

\item{template_type}{Type of template ("generic", "mortality", "labor")}
}
\description{
Creates a template configuration file that users can customize
}
