% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_generation.R
\name{create_f2_mortality_table}
\alias{create_f2_mortality_table}
\title{Create F2-style mortality damage projection table}
\usage{
create_f2_mortality_table(results_median, raw_data_path, tables_dir, config)
}
\arguments{
\item{results_median}{Median gamma coefficients (scaled)}

\item{raw_data_path}{Path to raw mortality data}

\item{tables_dir}{Output directory}

\item{config}{Configuration list}
}
\description{
Creates damage projections table similar to original f2_tables.R
}
