% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_management.R
\name{merge_with_defaults}
\alias{merge_with_defaults}
\title{Merge user config with defaults}
\usage{
merge_with_defaults(user_config)
}
\arguments{
\item{user_config}{User configuration list}
}
\value{
Merged configuration
}
\description{
Merges user configuration with default settings
}
