% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{process_damage_data}
\alias{process_damage_data}
\title{Process damage function data}
\usage{
process_damage_data(data, config)
}
\arguments{
\item{data}{Raw data from load_damage_data()}

\item{config}{Configuration list}
}
\value{
Processed data ready for analysis
}
\description{
Processes data following the reference code methodology but with flexible impact variable
}
