% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_components.R
\name{run_global_analysis}
\alias{run_global_analysis}
\title{Run global analysis}
\usage{
run_global_analysis(data, gamma, use_weights)
}
\arguments{
\item{data}{Processed data}

\item{gamma}{Gamma statistics}

\item{use_weights}{Whether to use population weights}
}
\value{
Global analysis results
}
\description{
Performs population-weighted global analysis
}
